Attempt to fix transition flickering

+++ engine/src/menus/menumanager.cpp
@@ -40,13 +40,13 @@
 
 //! This defines the list of events for the transition show animation
 const FrameEvent transition_show_event[] = {
-    {(uint16)-1, msc::NO_TRACK, NO_SOUND, 0x0, NULL }
+    {-1, msc::NO_TRACK, NO_SOUND, 0x0, NULL }
 };
 
 //! This defines the list of events for the transition leave animation
 const FrameEvent transition_leave_event[] = {
     { 1, msc::NO_TRACK, MENU_CHANGE,   0x1, NULL },
-    {(uint16)-1, msc::NO_TRACK, NO_SOUND, 0x0, NULL }
+    {-1, msc::NO_TRACK, NO_SOUND, 0x0, NULL }
 };
 
 /*!
